python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "faber"
version = "1.0.0"
description = "FABER - AI-assisted development workflows powered by LangGraph"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Fractary", email = "dev@fractary.com" }
]
keywords = ["ai", "agents", "workflow", "langgraph", "langchain", "development"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    "langsmith>=0.1.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.0", "langgraph-checkpoint-postgres>=2.0.0"]
redis = ["redis>=5.0.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
]
all = ["faber[postgres,redis,dev]"]

[project.scripts]
# faber = "faber.cli.main:app"  # REMOVED in v1.0.0 - Use @fractary/cli instead

[project.urls]
Homepage = "https://github.com/fractary/faber"
Documentation = "https://github.com/fractary/faber#readme"
Repository = "https://github.com/fractary/faber"
Issues = "https://github.com/fractary/faber/issues"

[tool.hatch.build.targets.wheel]
packages = ["faber"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
